Callbacks and functions can now be unregistered
In addition, unloading funcitonality for plugins is now active. This
means any registration of server resources in the entry of the plugin
MUST be unloaded at plugin unloading (any register_callback or
register_function must be matched by an unregister_callback or
unregister_function). In general, any resources attained in the entry
function must be released properly in the exit function.